更新检查点逻辑以正确保存最佳模型的epoch，启用基准测试，修正数据集长度计算，调整图像加载路径，增强数据增强返回值类型，并在命令行参数中添加训练相关选项

+++ SPEN/data/augmentation.py
@@ -145,7 +145,7 @@ class PerspectiveAug():
         ori_warpped = rotation * R.from_quat(ori, scalar_first=True)
         ori_warpped = ori_warpped.as_quat(canonical=True, scalar_first=True)
 
-        return image_warpped, pos_warpped, ori_warpped, box_warpped
+        return image_warpped, pos_warpped.astype(np.float32), ori_warpped.astype(np.float32), box_warpped.astype(np.float32)
 
 
 class ZAxisRotation():
@@ -199,7 +199,7 @@ class ZAxisRotation():
         ori_warpped = rotation * R.from_quat(ori, scalar_first=True)
         ori_warpped = ori_warpped.as_quat(canonical=True, scalar_first=True)
 
-        return image_warpped, pos_warpped, ori_warpped, box_warpped
+        return image_warpped, pos_warpped.astype(np.float32), ori_warpped.astype(np.float32), box_warpped.astype(np.float32)
 
 class AlbumentationAug():
     """
